fix+perf+chore: self-host CI from 41 min to 30 s, drop benchmarks and Zig
One commit, not six. The compiler sources, the regenerated canonical C, the
32 parity goldens and PUBLISHED_C have to move together, and PUBLISHED_C
lives in scripts/verify_seed.py, which also carries the Zig purge and the
compile cache. Any split needs partial staging of the same file across
commits, so the change lands as one reviewable unit instead.

Measured on AMD EPYC 7763, 2 cores available, 7.9 GB, gcc 13.3.0, Linux.
"Before" is the parent commit in a separate worktree with the compile cache
disabled; "after" is this tree. Wall clock, repeated runs.

  python scripts/build_selfhost.py --cc gcc --check-stale   205s ->  34s
  python scripts/verify_seed.py --cc gcc                     285s ->  10s
                                                              (27s cold)
  stress-gate, 8 legs in parallel          2485s serial  ->   ~27s

Where the 41 minutes went, and what moved it:

  61 compilations of the same 3.9 MB C unit, serially, with no cache. Six of
  the seven inside verify_seed.py produce byte-identical C at the fixed
  point, and orb_build recompiled with our own cc what the compiler had
  already compiled. The 8 stress iterations ran in a sequential for loop.

Now: a content-addressed compile cache (scripts/orbit_ccache.py, keyed on
compiler identity, version, exact flags and the SHA-256 of every input, so a
stale entry can only cause a miss); -Wall dropped from the bootstrap (5.7x on
that unit for 89 warnings, ORBIT_BOOTSTRAP_WARNINGS=1 restores it);
ORBIT_SKIP_CC so the compiler emits C without running a redundant cc; the 8
iterations as a CI matrix with an explicit needs:; actions/cache for the
compile cache. scripts/werror_gate.py, the strict -Werror gate that
ENGINEERING.md relies on but which was never wired into CI, is now a step.

The fixed point gate got stricter, not looser. verify_seed.py asserted the
emitted C of the seed only, so a chain that converged at stage one and
drifted afterwards passed. It now asserts chain2 and chain3 too, and when the
cache serves a contract binary the binary comparison reports "not evaluated"
rather than a tautological pass.

Compiler and runtime, by measured before/after:

  compiler/optimizer.orb  sinkOne was O(n^2) per call inside a 256-round
    driver, rebuilding the whole instruction list on each of its 127 motions
    (291,822 list pushes). One O(n) sweep per round over per-function side
    tables, motions in place. 20.3% -> 1.0% of self time, list traffic 64M ->
    under 1M. All 129 motions byte-identical to the old pass on 442
    functions.
  compiler/c_backend.orb  variableDefType rescanned every instruction for
    every name lookup. One O(n) index per function: 32,529,110 list reads ->
    40,907. The index caches the lookup, not the answer, because
    inferRegisterTypes rewrites register types as it runs. Emitted C proven
    byte-identical across 52 example, suite and std programs.
  compiler/fmt.orb, doctor.orb  21 quadratic `out = out + ...` sites and one
    arena allocation per character. Routed through the chunk buffer the C
    backend already used. `fmt --check compiler/c_backend.orb` went from
    SIGTERM at 2866 MB to 6.6 MB, and the whole compiler tree from 4.36 GB to
    64 MB, so the CI formatter gate no longer depends on runner RAM. Output
    byte-identical on 139 of 140 files, the 140th being the one the old
    compiler could not process at all.
  compiler/pipeline.orb  honours TMPDIR and TMP as well as TEMP, and a new
    orbit_os_temp_dir_selfhost extern means a plain `orbit build` no longer
    drops a multi-megabyte C file in the working directory.
  runtime/pulse.c  %llu with a uint64_t argument, via <inttypes.h> PRIu64.
    A real LP64 portability bug, found the moment the strict warning gate was
    actually run: correct where uint64_t is unsigned long long, broken where
    it is unsigned long.
  runtime/thread_pool.c  getsockname took int* where it wants socklen_t*.
    Signedness mismatch, now gone.

Windows CI, which failed with "URL can't contain control characters
'/Program Files/Git/gate-burst'". Git Bash is an MSYS2 host and rewrites any
argument that looks like an absolute POSIX path into a Windows path before a
native program sees it, so --burst-path /gate-burst arrived mangled. Phase A
passed because /health is the script's default and never a shell argument.
Fixed in two layers: scripts/orbit_routes.py repairs the value and rejects
what it cannot repair, wired into all three tools that take a route, plus
MSYS2_ARG_CONV_EXCL and MSYS_NO_PATHCONV at job level. scripts/routes_probe.py
(72 cases) is a CI step. Verified live against blog_api with the exact string
that failed, and with the three other spellings.

Removals. The benchmark suite (41 files), its methodology guide,
runtime/benchmark_arena.c, scripts/bench-energy.py and
scripts/bench-ledger-overhead.c, and every mention across 25 other files.
`grep -ri bench` and `grep -ri zig` both return nothing. The Zig purge covers
the `zig cc` fallback, the whole --bootstrap/--driver lineage in
verify_seed.py, the zig-out lookup in the VS Code extension, the dead
tests/parity/run_parity.{sh,ps1} harness, and all prose including
ENGINEERING.md and docs/ARCHITECTURE.md, which described a src/ tree that has
not existed for a long time. Consequence, documented in SOVEREIGNTY.md:
build_selfhost.py --promote is now the only way to replace the root of trust.

docs/PERF.md and docs/ENERGY.md keep their method, machine, rejected-
optimization reasoning and the wire-format bug the gate caught, and lose the
figures whose instrument no longer exists, labelled as needing re-measurement.

Documentation. docs/SYNTAX_GUIDE.md was a 5-line stub; its 11 samples are now
extracted, compiled and run, which corrected three that did not work: the real
union syntax, result needing a bound val because a bare `return ok(...)`
parses as a route response, and route programs getting main generated for
them. Fixed .gitattributes pointing at src/runtime/vendor/** and
examples/README.md documenting system.latency_ms() when the builtin is
system.latency_avg_us(). No broken markdown links. Added CODEOWNERS, issue
and PR templates, and the full gate list in CONTRIBUTING.md.

Three defects found on the way, recorded in ENGINEERING.md as debt rather
than worked around:

  CODE-0 (High)  type inference reads the previous function's string-typed
    variable names, so an unrelated change can alter the emitted C for an
    unrelated program. This is why "emitted C is unchanged" has to be proven
    by diffing and not by inspection.
  FMT-0 (High)   under memory pressure `orbit fmt <file>` truncates the file
    to 0 bytes and still exits 0. The CI formatter gate could have emptied a
    source file and reported success.
  CODE-1 (Medium) a user function named read, write, open or close collides
    with the C library declaration and fails at link time even though
    `orbit check` passes.

Not fixed, and not claimed: 92 -Wint-conversion warnings remain on the emitted
C, all one kind, because the register machine represents a value as an integer
and OrbitResult.value is void*. Root cause understood, tracked as STAB-3.
Making -Wall opt-in in the bootstrap is done; making the emitted C fully
-Werror clean is not.

Gates, on the compiler CI builds with verify_seed --emit-fixed-point:
build_selfhost --check-stale, verify_seed --release 5/5, parity 32/32,
fmt --check compiler, fmt --check tests/suite, doctor tests/suite, cli_probe
36/36, werror_gate 9/9, test_suite 25/25, test_suite --dir tests/std 13/13,
routes_probe 72/72, fuzz_frontend 150/150, the arena, http-parse, kynx and
auth C tests, and the live Kynx route-limit gate.

+++ runtime/pulse.c
@@ -14,6 +14,7 @@
 #include "kynx.c"
 #include "arena_pool.c"
 #include <stdio.h>
+#include <inttypes.h>
 
 /* ──────────────────────────────────────────────────────────────────────
  * Orbit Pulse — Real-time telemetry engine.
@@ -34,20 +35,24 @@ orbit_string orbit_pulse_get_stats_json(OrbitArena* arena) {
     
     // We are basically formatting a JSON string into the arena
     // In a real implementation we'd use a JSON builder, but a clean snprintf is fine for fixed schema
+    // PRIu64/PRIu32 (not %llu/%u) so the format always matches uint64_t/uint32_t:
+    // on LP64 targets uint64_t is `unsigned long`, which %llu does not match.
     char* buf = (char*)orbit_alloc(arena, 4096);
     snprintf(buf, 4096,
         "{"
-        "\"requests\":{\"total\":%llu,\"avg_cycles\":%llu,\"min\":%llu,\"max\":%llu},"
-        "\"memory\":{\"reuse\":%llu,\"alloc_mb\":%.2f,\"active_arenas\":%u},"
-        "\"security\":{\"blocks\":%llu,\"tracked_ips\":%u},"
-        "\"database\":{\"queries\":%llu,\"total_cycles\":%llu},"
-        "\"interning\":{\"hits\":%llu}"
+        "\"requests\":{\"total\":%" PRIu64 ",\"avg_cycles\":%" PRIu64 ",\"min\":%" PRIu64 ",\"max\":%" PRIu64 "},"
+        "\"memory\":{\"reuse\":%" PRIu64 ",\"alloc_mb\":%.2f,\"active_arenas\":%" PRIu32 "},"
+        "\"security\":{\"blocks\":%" PRIu64 ",\"tracked_ips\":%" PRIu32 "},"
+        "\"database\":{\"queries\":%" PRIu64 ",\"total_cycles\":%" PRIu64 "},"
+        "\"interning\":{\"hits\":%" PRIu64 "}"
         "}",
-        stats.request_count, avg_cycles, stats.min_cycles, stats.max_cycles,
-        stats.arena_reuse_count, (double)stats.total_alloc_bytes / (1024.0 * 1024.0), stats.active_arenas,
-        stats.kynx_blocks, stats.kynx_tracked_ips,
-        stats.db_queries, stats.db_total_cycles,
-        stats.string_intern_hits
+        (uint64_t)stats.request_count, (uint64_t)avg_cycles,
+        (uint64_t)stats.min_cycles, (uint64_t)stats.max_cycles,
+        (uint64_t)stats.arena_reuse_count, (double)stats.total_alloc_bytes / (1024.0 * 1024.0),
+        (uint32_t)stats.active_arenas,
+        (uint64_t)stats.kynx_blocks, (uint32_t)stats.kynx_tracked_ips,
+        (uint64_t)stats.db_queries, (uint64_t)stats.db_total_cycles,
+        (uint64_t)stats.string_intern_hits
     );
     
     return buf;
